// src/types/partials/FamilyMemberHistoryCondition.ts
/* eslint-disable no-unused-vars */
/* eslint-disable no-trailing-spaces */
// This file is auto-generated by generate_classes so do not edit manually

import { TExtension } from '../partials/Extension';
import { TCodeableConcept } from '../partials/CodeableConcept';
import { TQuantity } from '../partials/Quantity';
import { TRange } from '../partials/Range';
import { TPeriod } from '../partials/Period';
import { TAnnotation } from '../partials/Annotation';

export type TFamilyMemberHistoryCondition = {
    id?: String;
    extension?: TExtension[];
    modifierExtension?: TExtension[];
    code: TCodeableConcept;
    outcome?: TCodeableConcept;
    contributedToDeath?: Boolean;
    onsetAge?: TQuantity;
    onsetRange?: TRange;
    onsetPeriod?: TPeriod;
    onsetString?: String;
    note?: TAnnotation[];
};